Added tests for context processor and splitted tests.
